enable few more unit tests.
